Fix Flow warnings about TequilapiClient mocks
