+++ test/unit/renderer/store/modules/connection.spec.js
@@ -21,7 +21,6 @@ import {expect} from 'chai'
 import type from '@/store/types'
 import {mutations, actionsFactory} from '@/store/modules/connection'
 import {describe, it, beforeEach} from '../../../../helpers/dependencies'
-import factoryTequilapiManipulator from './tequilapi-manipulator'
 import {FunctionLooper} from '@/../libraries/functionLooper'
 import ConnectionStatusEnum from '../../../../../src/libraries/mysterium-tequilapi/dto/connection-status-enum'
 import communication from '@/../app/communication/messages'
@@ -32,6 +31,115 @@ import type {EventFactory as StatsEventsFactory} from '../../../../../src/app/st
 import {ActionLooper, ActionLooperConfig} from '../../../../../src/renderer/store/modules/connection'
 import ConnectionStatisticsDTO from '../../../../../src/libraries/mysterium-tequilapi/dto/connection-statistics'
 import type {BugReporter} from '../../../../../src/app/bug-reporting/interface'
+import TequilapiManipulator from './tequilapi-manipulator'
+import ConnectionStatusDTO from '../../../../../src/libraries/mysterium-tequilapi/dto/connection-status'
+import ConnectionIPDTO from '../../../../../src/libraries/mysterium-tequilapi/dto/connection-ip'
+
+function factoryTequilapiManipulator () {
+  let statusFail = false
+  let statisticsFail = false
+  let ipFail = false
+  let ipTimeout = false
+  let connectFail = false
+  let connectFailClosedRequest = false
+
+  const errorMock = new Error('Mock error')
+  const timeoutErrorMock = createMockTimeoutError()
+  const closedRequestErrorMock = createMockRequestClosedError()
+
+  class ConnectionApiManipulator extends TequilapiManipulator {
+    async connectionCreate (): Promise<ConnectionStatusDTO> {
+      if (connectFailClosedRequest) {
+        throw closedRequestErrorMock
+      }
+      if (connectFail) {
+        throw errorMock
+      }
+      return new ConnectionStatusDTO({})
+    }
+
+    async connectionStatus (): Promise<ConnectionStatusDTO> {
+      if (statusFail) {
+        throw errorMock
+      }
+      return new ConnectionStatusDTO({
+        status: 'mock status'
+      })
+    }
+
+    async connectionCancel (): Promise<ConnectionStatusDTO> {
+      return new ConnectionStatusDTO({})
+    }
+
+    async connectionIP (): Promise<ConnectionIPDTO> {
+      if (ipTimeout) {
+        throw timeoutErrorMock
+      }
+      if (ipFail) {
+        throw errorMock
+      }
+      return new ConnectionIPDTO({
+        ip: 'mock ip'
+      })
+    }
+
+    async connectionStatistics (): Promise<ConnectionStatisticsDTO> {
+      if (statisticsFail) {
+        throw errorMock
+      }
+      return new ConnectionStatisticsDTO({duration: 1})
+    }
+  }
+
+  return {
+    getFakeApi: function () {
+      return new ConnectionApiManipulator()
+    },
+    cleanup: function () {
+      this.setStatusFail(false)
+      this.setStatisticsFail(false)
+      this.setIpFail(false)
+      this.setIpTimeout(false)
+      this.setConnectFail(false)
+      this.setConnectFailClosedRequest(false)
+    },
+    setStatusFail: function (value: boolean) {
+      statusFail = value
+    },
+    setStatisticsFail: function (value: boolean) {
+      statisticsFail = value
+    },
+    setIpTimeout: function (value: boolean) {
+      ipTimeout = value
+    },
+    setIpFail: function (value: boolean) {
+      ipFail = value
+    },
+    setConnectFail: function (value: boolean) {
+      connectFail = value
+    },
+    setConnectFailClosedRequest: function (value: boolean) {
+      connectFailClosedRequest = value
+    },
+    getFakeError: function (): Error {
+      return errorMock
+    }
+  }
+}
+
+function createMockTimeoutError (): Object {
+  const error = new Error('Mock timeout error')
+  const object = (error: Object)
+  object.code = 'ECONNABORTED'
+  return error
+}
+
+function createMockRequestClosedError (): Object {
+  const error = new Error('Mock closed request error')
+  const object = (error: Object)
+  object.response = { status: 499 }
+  return error
+}
 
 const fakeTequilapi = factoryTequilapiManipulator()
 const fakeMessageBus = new FakeMessageBus()
